+++ packages/swarm-components/src/index.js
@@ -4,6 +4,7 @@ export { default as Checkbox } from './Checkbox';
 export { default as Icon } from './Icon';
 export { default as Radio } from './Radio';
 export { default as TextInput } from './TextInput';
+export { default as Textarea } from './Textarea';
 export { default as Toggle } from './Toggle';
 export { default as TogglePill } from './TogglePill';
 export { default as Select } from './Select';